backend/requirements.txt
fastapi
uvicorn
pandas
numpy
faiss-cpu
sentence-transformers
langchain
requests
python-dotenv
uvicorn